+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v1.4.1 - Table resizing & new Setting
+
+### Added
+
+- Settings
+  - Update collapsed Folders: How long the delay between updates to the master property "collapsedFolders" is.
+- Tables can now be resized
+
 ## v.1.4.0 - Folder Collapsing & Improvements
 
 ### Added